+++ lapce-app/src/doc.rs
@@ -2177,6 +2177,7 @@ fn should_blink(
                 | Focus::Panel(PanelKind::Plugin)
                 | Focus::Panel(PanelKind::Search)
                 | Focus::Panel(PanelKind::SourceControl)
+                | Focus::Panel(PanelKind::Agent)
         ) {
             return true;
         }